+++ backend/canisters/community/impl/src/updates/join_channel.rs
@@ -61,7 +61,9 @@ fn is_permitted_to_join(
 
             if let Some(channel_member) = channel.chat.members.get(&member.user_id) {
                 Err(AlreadyInChannel(Box::new(
-                    channel.summary(Some(channel_member.user_id), state.env.now()).unwrap(),
+                    channel
+                        .summary(Some(channel_member.user_id), state.data.is_public, state.env.now())
+                        .unwrap(),
                 )))
             } else {
                 Ok(channel
@@ -84,12 +86,12 @@ fn commit(channel_id: ChannelId, user_principal: Principal, state: &mut RuntimeS
             let now = state.env.now();
             match join_channel_unchecked(channel, member, state.data.is_public, now) {
                 AddResult::Success(_) => {
-                    let summary = channel.summary(Some(member.user_id), now).unwrap();
+                    let summary = channel.summary(Some(member.user_id), state.data.is_public, now).unwrap();
                     handle_activity_notification(state);
                     Success(Box::new(summary))
                 }
                 AddResult::AlreadyInGroup => {
-                    let summary = channel.summary(Some(member.user_id), now).unwrap();
+                    let summary = channel.summary(Some(member.user_id), state.data.is_public, now).unwrap();
                     AlreadyInChannel(Box::new(summary))
                 }
                 AddResult::Blocked => UserBlocked,